requirement.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
schedule==1.2.0
aiohttp==3.8.5
sqlite3